requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
google-generativeai==0.3.2
python-dotenv==1.0.0